lookup.c: do not warn about deprecated functions
This is a "tactical" fix: we are about to merge a large
extension of this module (actually the first complete implementation)
what probably fixes the root problem anyways.

see https://github.com/rsyslog/rsyslog/pull/578

After we have merged, we need to re-enable those warnings. For
now, that saves us trouble with travis (and, no, I have no idea why
the warnings did not break travis so far...).