imuxsock: fix segfault when processing system log socket
The root cause of this problem is that in commit
16db662d9fac0f9636eea873d690a629641be5bc rulesets where
introduced. However, the system log socket ruleset pointer
was not properly initialized. So it depends on the memory state
which value the ruleset pointer has.

Note that this problem occurs only when running with sufficient
privileges to read the root log socket AND memory is not zeroed-out.
Under these conditions, a segfault rather quickly on startup is
possible, but does not always happen. As it is an uninitialized
data bug, problems can also occur any time after startup. Besides
a segfault, message may also be lost due to assignment to invalid
ruleset.

This regression was caused by an update post the 8.16.0 release
and is fixed before 8.17.0, so no officially released version
is affected.