build system: fix build issues with clang
clang builds often failed with a missing external symbol
"rpl_malloc". This was caused by checks in configure.ac,
which checked for specific GNU semantics. As we do not need
them (we never ask malloc for zero bytes), we can safely
remove the macros.

Note that we routinely run clang static analyer in CI and
it also detects such calls as invalid.

closes https://github.com/rsyslog/rsyslog/issues/834