omfwd bugfix: keepalive action parameters did not work
due to being inconsistently spelled. This lead to an internal
program error, which unfortunately was only reported to the
debug log.

This patch both fixes the problem and also emits the internal
program error message to the regular log stream (in case a similar
problem re-occurs in the future). A testbench test is also added.

closes https://github.com/rsyslog/rsyslog/issues/916