Add offset to metadata for msgs from imfile
The offset of the beginning of the log line is now added as metadata, if
additional metadata is requested.

There is one odd case, where a partial line is read from a file an EOF
is hit before the proper end of the msg is encountered (for any mode or
multiline msgs).  In that case, the offset of the remaining record in
the subsequent file is used for the entire msg.

We also take this opportunity to consistently use off64_t instead of a
combination of int64, off_t, and off64_t (allowing consistent behavior
in 32 bit environments).