bugfix: memleak in worker thread termination
Template buffer was not freed for actions which did at least
one of the following

- did not create instances
- were non-transactional

This was a relatively small leak. However, it happened each time
the worker was shut down. So in environments where workers start
and shut down continously, this could consume a larger amount of
memory after some time.

closes https://github.com/rsyslog/rsyslog/issues/968