testbench fix: execonlywhenprevsuspended_multiwrkr.sh occasionally failed
The root cause of the problem was a race condition. Test instrumentation
sets failure state after 2 messages. It expected that these messages were
actually the first ones injected. Especially on slower systems, a second
worker may begin processing before the first one, in which case the first
two messages have different numbers. We have changed this to ensure that
first two are always those that we expect.

Note that this does not compromise the test: all worked well, it was just
a race in regard to the test predicate we check. So actually the test
setup was wrong, not the test method as such or the to-be-tested code.

closes https://github.com/rsyslog/rsyslog/issues/842