testbench: detect rsyslog abort in startup phase
A segfault in startup can happen with some bugs. This required the testbench
to timeout waiting on startup files. While this is OK, it defers testbench
progress, and could cause to exceed CI systems runtime limit. That in turn
lead to loss of diagnostic information (because test-suite.log is not written).

This commit offers a quicker abort detection.

closes https://github.com/rsyslog/rsyslog/issues/1000