pmrfc3164: Add support for at-signs in hostname
This commit adds support for new parameter permit.AtSignsInHostname
to allow at-signs characters in hostname rfc3164 BSD field. The parameter
value default to 0. The patch is similar to the one introduced by
Rainer Gerhards <rgerhards@adiscon.com> in 8.20.0 to support slashes
characters in hostnames.

Slashes and at-signs may be inserted by syslog-ng at least on 3.2 when
it runs with use_dns(no), keep_hostname(no) and chain_hostnames(yes).
The at-sign may happen when the syslog-ng host sends local logs in rfc3164
format. In that senario, the hostname field is prefixed by the source
name and an at-sign character.
As for the slash character, it is inserted by syslog-ng to chain
hostnames and ip addresses accross relays.