testbench: add test for imfile "reopenOnTruncate" mode
The problem checked with this test has now been fixed.
This activates the test inside the testbench to guard against
future regressions. It also changes the test, because the old
construct caused trucation and immediate rewrite with the same
nubmer of bytes, so that the trucation could not be detected.