bugfix: potential hang on rsyslog startup
When we log internal messages to syslog (the default starting with 8.21),
early error messages can result in rsyslogd hanging, as well as everything
else that wants to log hanging.

This is the result of deadlock: rsyslog tries to write to the log socket,
which blocks but can't be unblocked as rsyslog processing would need to
advance to do so. Interestingly, this only happens occasionally (even
though the startup sequence is deterministic at this point), so it is
hard to diagnose. Also, not all platforms seem to be affected. Platforms
with systemd journal will never run into a deadlock (as journal in this
case reads the system log socket).

The root cause is that for internal messages we maintain a queue for
such early messages and emit them only after full startup. Similar code
is missing when writing to the log socket. The full code has been
restructured so that this condition is obeyed under all conditions.

closes https://github.com/rsyslog/rsyslog/issues/1188