imuxsock: do not drop messages from our own pid if we log to syslog()
On non-systemd journal systems, rsyslog messages generated from the
same instance that reads the system log socket are ignored by default.
This is because imuxsock by default ignores messages from its own
process. This makes sense if messages are internally processed,
but does not make sense if the shall go via the syslog call.

This is a longer standing issue, but became more important starting
with 8.21, which logs to syslog by default.

closes https://github.com/rsyslog/rsyslog/issues/1194