bugfix: segfault if hostname is unset on system
happens when gethostname() returns empty string. This will cause
the createon of the localhostname prop_t to fail, which in turn
leads to a NULL pointer dereference when emitting local messages.
As we emit a startup message by default, this had usually lead
to a segfault on startup.

Thanks to Erik Potter and github user mostolog for their help
in analyzing this problem.

closes https://github.com/rsyslog/rsyslog/issues/1040
closes https://github.com/rsyslog/rsyslog/issues/335