core: fix sequence error in msg object deserializer
Corruption of disk queue (or disk part of DA queue) always happens if
the "json" property (message variables) is present and "structured-data"
property is also present. This causes rsyslog to serialize to the
queue in wrong property sequence, which will lead to error -2308 on
deserialization.

Seems to be a long-standing bug. Depending on version used, some or
all messages in disk queue may be lost.

closes https://github.com/rsyslog/rsyslog/issues/1404