iminteral: add addtl debug message
We have seen cases where the kill() used to signal new messages
actually cases problems and the new debug message is intended to
provide a better clue to what is happening. In any case, this
is generally useful, and so should remain even after the debugging
of that special case.