apply consistent code style as defined in .clang-format file
Note: this commit does not actually change any code except
for style reasons. Unfortunately, the style change requires
many changes all over the place. As such, diffs of versions
before this patch and after are almost unreadable. We accept
this fate as we otherwise cannot get to a consistent code style.

HOWEVER, git bisect will continue to work well, and is highly
recommended when hunting for regressions.