bugfix: handling of json "null" objects in regard to msgGetJSONPropJSON
libfastjson returns NULL for json "null" objects, and callers need
to handle this. This is correctly handled in almost all code path
in rsyslog, and this patch fixes theoretical issues in the remaining
two places. Note, however, that a json null object can not really
occur at those places due to the rest of processing. But we
wanted to include the check anyways.

closes https://github.com/rsyslog/rsyslog/issues/600