omkafka: Implemented newer delivery callback from librdkafka.
This somehow elimitated a lot of failing messages which were not
processed on shutdown (flush) before. The old callback API was
depreceated.

This should also elimitate most of the message loss when
kafka queue is full from issue:
https://github.com/rsyslog/rsyslog/issues/1052

The new code was tested with
queue.buffering.max.messages=100 and 2000 injected messages.

Testbench: Try starting kafka server second time if failed on
first try. Sometimes the broker is still registered in zookeeper.