imfile: multiline timeout does not work after rsyslog restart
The timeout feature for multiline reads does not correctly work
for files for which a state file existed. This is usually the
case for files that had been processed by a previous run and
that still exist on the new start. For all other files,
especially those monitored by a wildcard and newly created after
the rsyslog start, timeout worked as expected.

closes https://github.com/rsyslog/rsyslog/issues/1445