testbench bugfix: hang in tests if omstdout is not present
Many tests depend on omstdout. Given the fact that omstdout
is really only useful for the testbench (at least that's the intent),
we now require --enable-omstdout if --enable-testbench is given.

The alternative would have been to disable all those tests that
need it, which would have lead to considerable less testbench
coverage.

closes https://github.com/rsyslog/rsyslog/issues/1649