omelasticsearch: fixing leak while writing error file
since we switched to libfastjson, the default error function is
duplicating the string from fjson_object_to_json_string in order to free
the fjson object.
we however don't free the duplicated string later

fix that by adding a free
also add a second duplication for the second case of rendered string;
this makes the code easier to read instead of adding another condition.

Fixes: issue #1668
Signed-off-by: William Dauchy <w.dauchy@criteo.com>