Fix Elasticsearch response handling
Elasticsearch responses for bulk requests differ slightly between
versions <= 2.4 and more recent ones (5 and upwards).

As a matter of fact, they used to look like this for version 2.4 :
'{
    "took":1205,
    "errors":false,
    "items":[{
        "create":{
            "_index":"log-2017.08.25",
            "_type":"log",
            "_id":"AV4JLGyt91GzIgdyx6s7",
            "_version":1,
            "_shards":{
                "total":2,
                "successful":1,
                "failed":0
            },
            "status":201
        }
    }]
}'

Whereas they now look like this, starting from version 5 (there was no
version between 2.4 and 5):

'{
    "took":1829,
    "errors":false,
    "items":[{
        "index":{
            "_index":"log-2017.08.25",
            "_type":"log",
            "_id":"AV4JPZ_fquWCvYqrs7tN",
            "_version":1,
            "result":"created",
            "_shards":{
                "total":2,
                "successful":1,
                "failed":0
            },
            "created":true,
            "status":201
        }
    }]
}'

Note that the main difference is that the object "create" was renamed as
"index".

So far, the code was looking for the key "status" of this object
"create". This commit make it so that it now also tries to find the
object "index" if the object "create" does not exist.

Fixes #1731