omfwd bugfix: UDP oversize message not properly handled
When a message larger than supported by the UDP stack is to be sent,
EMSGSIZE is returned, but not specifically handled. That in turn
will lead to action suspension. However, this does not make sense
because messages over the UDP max message size simply cannot be sent.

This patch truncates the message in question and reports an error
message with the condition.

closes https://github.com/rsyslog/rsyslog/issues/1654