core: double free under OOM condition (extremely unlikely)
Detected by Coverty scan, can only happen when we run out of
memory, in which case we have a very big problem anyhow.

Double free could happen in lower layer when a property was
freed due to realloc failure. Solution is to continue to use
old value, which also makes sense otherwise.