omstdout: fix theoretical buffer overrun
The overrung occurs in code for testing an interface that currently
does not exists, so the overrun can actually never happen. We fixed
it anyhow to keep Coverity scan clean of defects.

Note also that omstdout is a testbench tool not intended for
production.

The bug was originally tried to solve in
a3326ed49d9a4908342e3e419dc942a5e2895afb, but that commit was
incomplete.

Detected by Coverity scan, CID 185325