remove systemd embedded code, use library instead
Since the early days rsyslog uses the orginal systemd embedded
interface code. This commit changes it to libsystemd where available.
If we do not find libsystemd, we assume the system does not use
systemd, which is a safe assumption nowadays. This ensures we use the
fresh interface code and also removes bloat from our project.

closes https://github.com/rsyslog/rsyslog/issues/1933