imjournal: refactor journal state file load
Remove "time of check time of use" Coverity scan warning by refactoring
the code to directly try out things prior to just checking if we could.
Ultimately, we need to handle the same error (and already did so), so
there was little point in doing a check first.

Coverity scan CID 185500