imptcp: check if we really got a remote addr on connect
CI testing seems to indicate that we sometimes do not get valid
data, but so far we could not reliably reproduce this. This commit
introduces an additional check which emits an error message in
that case. The hope is that we will see it on further occasions or
get some reports. If that really happens, we should handle it.